Optional page offset mode
This commit is build on the idea of Renato Florentino to be able to
correct the page offset that some documents have. His patch, where
most of the code has been adapted to this commit ( thanks! ), made
it able to correct the current page and automatically calculate the
correct page offset. Now this patch will introduce a third mode to
the bcmd_goto command: You are now able to set the page directly,
calculate it over the offset or search the labels for it.