Introduce a shortcut list
This commit introduces a shortcut list that is used
now to check given keybindgins: On the contrary to
the static shortcut array it can be easily modified
and extended.

In addition some malloc-checks have been introduced
and a named shortcut list that will be used to evaluate
the cmd_map parameters.