let the document know about the current adjustments
The document object now has functions to set and get the position and
the viewport size.

The position is a relative position with respect to the size of the
entire document, i.e. position_y=0 means top of the document and
position_y=1 bottom.

The viewport size is stored in screen coordinates, in pixels.