Created an example `zathurarc`
There is more that needs to be done (namely add some colour options).  I
feel that adding mouse bindings is unnecessary, but that is doable.  It
should cover every important keyboard feature.  It may actually
contradict the default settings (mainly with the presentation settings),
and it also sets copying to the system clipboard by default.

There is a TODO at the beginning of the script, but the only important
thing of note is that it needs to be better organised.