Show grid padding when recode (black pages)
It is nice to see the page grid both normally and when colors have been
recoded. The easy way to do this is to change the grid background to
gray instead of pure black, so it contrasts against either white pages
(typical of most documents) or black pages (typical recoded color).