Add '$f' and '$p' expansions to 'exec' shortcut function
There are other expansions available, but I don't really think they make
sense, or why there are two of them for the same thing (`$FILE` and `%`
expand to the file name).

I added `$f` and `$p`, which expand to the document's path and current
page respectively, while keeping the other 2 just in case it breaks
someone's setup. I also documented them in both `zathura(1)` and
`zathurarc(5)`.