utils: Add parser/writer for 'first-page-column'
The 'first-page-column' setting is a string representing a list of integers.
This is annoying to manipulate.  Currently, it has a single accessor,
'find_first_page_column', which looks up the nth value.
The next commit will add a function to change it.

To avoid this turning into a horrible buggy mess of string manipulation,
add functions to convert the settings string to and from integer arrays.
Rewrite 'find_first_page_column' using these functions.

Efficiency: the new functions are slower than 'find_first_page_column'
because they parse all values in the string.  This is irrelevant since
the string is supposed to be tiny, and is only used when opening a
document or when the page layout is modified by the user.