feat(render): optional lightness adjustment when recoloring
Some thinner fonts (e.g., Computer Modern Roman) can be harder to read
when recoloring. This implements the option recolor-adjust-lightness,
which adjusts the lightness before recoloring, improving readability.

The adjustment is implemented with a quartic curve. The adjusted
lightness is then averaged against the original lightness, weighed by
saturation, which makes colors less "washed out" (this was particularly
annoying for graphics like TikZ). The weighing is not implemented when
recolor-keephue is disabled for technical reasons, although it also
improves the readability of darker colors.