Add vim's expand/collapse fold bindings to index mode
The tree in the index mode is comparable to folds in vim, thus its
mappings around opening and closing folds make sense here too. Add the
following aliases:

- zo to expand
- zc to collapse
- zR to expand all
- zM to collapse all

and the following bindings to new functions:

- zO to expand recursively
- zC to collapse recursively

Additionally, do some minor reshuffling of the collapse* and expand*
entries in the touched code for a more consistent order.